refactor: DDD, SPR, SOILD
